claims: save claims and validate migration.
Refactored the spec format to better express alternatives. Specifically,
to nest fields under paths, so match alternatives can express "this
table with these fields or that table with those fields".

As a result, refactored a lot of spec data and claim setting machinery.

Also implemented the SQL functions to identify unmet claims and validate
the migration accordingly. This supports alternatives and optionals and
allows additive migrations to pass.