claims: regenerate on DDL update.
When the database changes, re-expand the spec and update the claims.

This works by the replication listener monitoring for inserts to a `:namespace.sync` table. When these occur, it loops through all of the specs and re-generates their claims.

This is not very efficient at the moment, as most of the time we brute
create the same claims. This implementation could be improved in `Integrate.Specification.sync_spec`.

In addition, the `integratedb_validate_migration()` function now calls
an `integratedb_sync()` when successful. This inserts the sync record
that then triggers the update.

Users can call integratedb_sync() directly, for example on rollback or
when deliberately skipping validation.